added day boxes with svg background and also Iframes
Added day boxes with a pattern that is created with svg. when the box is clicked it will increase in size. When clicked an event handler is adding a class to the box from which css will change the box to a fixed and centered position on the screen.
In this new location the Iframe within the div will now change it's visibility to visible and a new html file is presented where a connections game can be played.